refactor(cli): the resident line and the admission tests get their own files
Hygiene vector 12 (file-loc-cap) refused the push: main.rs read 1527
lines after the resident line (ADR-132) and server/tests.rs 1582 after
the admission tests (ADR-131). The resident line moves to
crates/nika-cli/src/resident.rs; the three admission tests move to
crates/nika-serve/src/server/admission_tests.rs; both files are back
under the cap with headroom. The ADR index is regenerated as the union
of this PR's ADRs and the sibling PR's (133 and 134), so both PRs carry
the identical generated index and the second merge is clean.
