+++ docs/adr/index.toml
@@ -1,7 +1,7 @@
 # ADR Index -- auto-generated by scripts/adr/generate-index.sh
 # Do not edit manually. Re-generate with: scripts/adr/generate-index.sh
 schema_version = "v1"
-count = 93
+count = 95
 
 [[adr]]
 id = "ADR-001"
@@ -920,4 +920,24 @@ filename = "adr-132-identity-idempotency-protocol.md"
 tags = ["architecture", "one-door", "idempotency", "protocol", "serve"]
 affects_crates = ["nika-serve", "nika-cli-host", "nika-cli"]
 affects_layers = ["L4"]
+
+[[adr]]
+id = "ADR-133"
+title = "The session machine is portable: typed outcomes, proposal and gate identities, hosts that judge by identity"
+status = "accepted"
+date = "2026-09-04"
+filename = "adr-133-the-session-machine-is-portable.md"
+tags = ["architecture", "session", "one-door", "idempotency"]
+affects_crates = ["nika-session", "nika-cli"]
+affects_layers = ["L4"]
+
+[[adr]]
+id = "ADR-134"
+title = "Trust is a rung, never a word: declared · discovered · observed on every access lane"
+status = "accepted"
+date = "2026-09-04"
+filename = "adr-134-trust-is-a-rung-never-a-word.md"
+tags = ["architecture", "access", "trust", "one-door"]
+affects_crates = ["nika-types", "nika-providers", "nika-service-execution", "nika-cli-host", "nika-cli"]
+affects_layers = ["L0", "L2", "L3", "L4"]
